requirements.txt
Flask==3.0.0
pyngrok==7.0.0
python-dotenv==1.0.0
